ARRRRRRRRRGenerate md report
So tired of CI systems... here's a super hack.

On App quit or test failure, generate a `.md` fle which has all of the log files from selenium stack, electron process logs, and a screenshot embedded as base64.  See app test for usage as a "crap... tests are about to fail...".